requirements.txt
psycopg2-binary == 2.8.4
python-dateutil == 2.8.1
pytz == 2019.3
PyYAML == 5.1.2
six == 1.13.0
pandas == 1.0.1
matplotlib == 3.1.3
seaborn == 0.10.0
tqdm == 4.43.0
ruts == 0.3.0
wordcloud == 1.6.0
scikit-learn == 0.22.2